requirements.txt
openai
python-docx
requests
python-dotenv
